refactoring and cleanup for landing page and app menu
